Refactor existing pipeline steps
