+++ tests/test_graphql_api/test_helsinki_profile/test_query.py
@@ -4,8 +4,8 @@ from typing import TYPE_CHECKING, Any
 
 import pytest
 
+from tilavarauspalvelu.enums import ADLoginAMR, LoginMethod, ProfileLoginAMR
 from tilavarauspalvelu.integrations.helauth.clients import HelsinkiProfileClient
-from tilavarauspalvelu.integrations.helauth.typing import ADLoginAMR, LoginMethod, ProfileLoginAMR
 from tilavarauspalvelu.integrations.sentry import SentryLogger
 
 from tests.factories import ApplicationFactory, ReservationFactory, UserFactory
