+++ tilavarauspalvelu/api/graphql/types/helsinki_profile/types.py
@@ -8,8 +8,9 @@ from graphql import GraphQLError
 from query_optimizer.selections import get_field_selections
 
 from tilavarauspalvelu.api.graphql.extensions import error_codes
+from tilavarauspalvelu.enums import LoginMethod
 from tilavarauspalvelu.integrations.helauth.clients import HelsinkiProfileClient
-from tilavarauspalvelu.integrations.helauth.typing import LoginMethod, UserProfileInfo
+from tilavarauspalvelu.integrations.helauth.typing import UserProfileInfo
 from tilavarauspalvelu.models import Application, Reservation
 from tilavarauspalvelu.tasks import save_personal_info_view_log
