+++ tilavarauspalvelu/integrations/helauth/utils.py
@@ -1,12 +1,18 @@
 from __future__ import annotations
 
 import base64
+import contextlib
 import json
-from typing import Any
+from typing import TYPE_CHECKING, Any
 
 __all__ = [
     "get_jwt_payload",
 ]
+
+
+if TYPE_CHECKING:
+    from tilavarauspalvelu.models import User
+    from tilavarauspalvelu.typing import WSGIRequest
 
 
 def get_jwt_payload(json_web_token: str) -> dict[str, Any]:
@@ -14,3 +20,17 @@ def get_jwt_payload(json_web_token: str) -> dict[str, Any]:
     payload_part += "=" * divmod(len(payload_part), 4)[1]  # Add padding to the payload if needed
     payload: str = base64.urlsafe_b64decode(payload_part).decode()  # Decode the payload
     return json.loads(payload)  # Return the payload as a dict
+
+
+@contextlib.contextmanager
+def use_request_user(request: WSGIRequest, user: User) -> None:
+    """
+    Use the provided user as the request user for the duration of the context.
+    This is needed during login, since the request user is still anonymous.
+    """
+    original_user = request.user
+    try:
+        request.user = user
+        yield
+    finally:
+        request.user = original_user
